factor out index getstaticprops
